:wrench: Update Rutford icequake example
Update the Rutford icequake example to slightly adjust the run
parameters and update the notebook to reflect these changes.

The lookup table was refined to focus on a smaller volume, with the
resolution increased. The sampling rate used to compute detect/locate
was adjusted to 250 Hz (down from 500/1000 Hz, respectively), since
using a lowpass filter of 124 Hz led to no measurable changes in the
final reported event location.