:wrench: Bump version number to 1.1.0rc1
Bumping version number for release candidate.

Note: Worth looking into whether it's possible to override the version
with a release candidate in some other way, so as to not have to change
the pyproject.toml file and add another commit once the release
candidate is passing with the test PyPI release test?