:wrench: Further tune icequake examples
Tune the icequake examples to use more appropriate marginal window
values, resulting in better locations with smaller uncertainties. In the
case of the Iceland icequake example, this also allowed for the minimum
event interval to be reduced and two additional events that preceded the
primary event focussed on in the example were found. The duration of the
marginal window was set be approximately the full duration half maximum
as determined from the trigger summary figures.

Benchmarks for the Iceland icequake example have been updated to reflect
these changes, with new benchmark files added for the two new events.