:wrench: Address future warning for pandas concat
Address future warning for behaviour of pandas concatenation in the case
of an empty DataFrame. We always had an empty DataFrame as we were
building it event by event, so the first event was always being
concatenated with an empty dataframe that just contained information
about the column names.

All three examples run the trigger stage with no complaints.