feat(trigger): Add median_ratio threshold method
Introduces a new method to determine the trigger threshold, by taking a
multiplier of the median value of the coalescence trace, in a
user-specified window duration.

In this commit the existing MAD multiplier method is not renamed from
"dynamic"; instead the new method is identified by an additional
threshold_method option "median_ratio". We might consider renaming from
"dynamic" to "mad_ratio" or similar to avoid ambiguity, but this would
be a breaking change.

Additional small tweaks surroundng this feature addition:
 - pass 'threshold_method' through to the trigger_summary() plotting
   function
 - add an example usage of the median_ratio threshold in the
   Volcanotectonic_Iceland example trigger() script
 - fix a bug in trigger.chunks2trace() caused by the deprecation of
   numpy.product() (--> np.prod() )