:test_tube: Add unit tests for trigger thresholds + smoothing
Add unit tests for each dynamic threshold method (mad and median ratio)
and for the trigger smoothing functionality. This includes moving the
nuts and bolts of the smoothing job to a new private function (motivated
by the test, but also makes things neater).

Test script has been added to the `run_tests.yml` for GitHub Actions.