:test_tube: Update test coverage process
Update the process for analysing package coverage of the tests. We were
encountering an issue where CodeCov could not properly parse the reports
because the paths in the .xml report were absolute paths to the
site-packages directory on the run machine, rather than relative to the
package top-level (`quakemigrate`).

Add a codecov.yml file that should adjust these paths when sending the
report. Also added the tests for specific parts of the package (under
`tests/`) to the coverage report.