:wrench: Always use "COA" timeseries for trigger peak index
After some discussion in #179, we have decided to revert the
change originally made in 7a5389c ; i.e. returning to always using the
raw "COA" timeseries to determine the timing of the coalescence peak
assigned to a candidate event within trigger. This is irrespective of
whether the "COA_N" timeseries has been used to set the trigger
threshold and identify peaks.

The motivation for this is that the "COA" timeseries reflects what is
used to determine the origin time in locate, and is therefore the most
direct comparison (given the relatively strict criteria on origin time
for the event to be "accepted").

Updates to affected benchmarks will follow.