:test_tube: Update benchmarks following `trigger_on` change
Update affected benchmarks following the change to always find the peak
index from the "COA" trace in trigger (836521a).

Only one event in the `Icequake_Iceland` example is affected (and only
in the trigger time / EventID; the locate output is identical -- and the
trigger time is now closer to the final origin time from locate).