:wrench:/:test_tube: Handle single mag estimate, update benchmarks
Add handling for case where there is only one valid magnitude estimate,
so that:
- the uncertainty estimate for that individual magnitude estimate is
  used to provide a magnitude uncertainty estimate for the event
- it is not attempted to calculate an r2 value (which leads to a
  divide-by-zero error, as a single datapoint has zero variance)

Also catch the case where all available amplitude estimates are
identical, and also skip the r2 calculation in this case.

Two events in the VT_Iceland benchmark are affected, as each only has a
single amplitude/magnitude estimate. They now have an uncertainty
estimate in the .event file, and print a message that the r2 estimate
will be skipped, rather than a RuntimWarning.