:atom: Add new Askja volcano VT+DLP example
Add a new example use-case, which uses data from around Askja volcano,
Iceland. This example (though expanded to run over a full day of data;
cut down here just to save on computation time for demonstrating the
code) is used in the manuscript to demonstrate the capability of
QuakeMigrate to detect and locate a wide variety of seismic events with
different source types, and in this case simultaneously with the same
set of parameters.

As in the `Volcanotectonic_Iceland` example, the waveforms are
downloaded before running detect/trigger/locate, so are not included in
the GitHub upload. Here we use the obspy MassDownloader.

More context about these events can be found in:
Greenfield, T. and White, R.S., 2015. Building Icelandic igneous crust by
repeated melt injections. Journal of Geophysical Research: Solid Earth,
120(11), pp.7771-7788.

and

Winder, T., 2022. Tectonic earthquake swarms in the Northern Volcanic
Zone, Iceland (Doctoral dissertation).