:wrench: Turn off pick plotting for Askja and VT_Iceland examples
It isn't typically worth the time taken to make all the pick plots,
unless a user has some specific reason to investigate them. They are
left turned on in the two icequake examples (where there are
comparitively few events) to demonstrate the capability (and as a place
to check for breaking changes in that part of the code), but disabled on
these two examples where they mis-prompt the user to make these plots,
and slow down the examples and tests based on them.