:wrench: Switch VT_Iceland example to MassDownloader
Switch the VT_Iceland example to use the obspy MassDownloader functions
for obtaining data from IRIS/EarthScope, instead of the simple client
interface. This is faster and cleaner.

The only slight downside is that if you re-run the MassDownloader with
the same target directory, it will generally not re-download the data if
you just slightly tweak the start and end times, as the test for whether
the data has been downloaded already is not very sophisticated. The
solution is to clear the target directories first.

Note also I have left `mdl.download.threads_per_client=3` (the default)
but this might need adjusting in case it causes any re-occurrence of #172.