:atom: Use channel maps to select components for waveform plotting
Previously the channel mapping for the data to be plotted in the phase
pick and event summary (waveform gather) plots was hard-coded (P=Z;
S=[N,1,E,2]), preventing waveforms being plotted if they did not conform
to this, or leading to data being plotted for the wrong phase if the
channel mapping was unusual.

Here a new utility function is added `get_phase_component_strings` to
generate regex strings for use within `Stream.select(component=<str>)`
from the `onset.channel_maps` variable set by the user. This is then
passed to the plotting functions and used to select the correct
components from the data to plot.

In doing so, we move one step closer to "phase agnosticism", but are
still bound to P and/or S phases only for the plotting -- this aspect is
still hard-coded but a different and bigger job to navigate. A
related limitation is the truncation of component lists after one pair
of alphabetical and/or numerical component codes for the S-phase (to fit
the model of two horizontal components feeding into the S onset
determination that is implicit in our plotting routines). Again, this
could be solved with some more imagination and application but is not
attempted here.

For P-phase, all components are retained and therefore plotted, and
adjustments have been made so that this will be reflected in the labels
- though waveforms will be overlain in the same colour, currently. This
should solve #158.

In addition " ({ph})" is appended to the waveform gather legend, to help
keep track of which component is feeding into which onset function when
the channel mapping gets wacky...