:atom: Use MassDownloader to get data for Icequake_Rutford
Switch from using waveform data shipped within the repo to using the
MassDownloader to efficiently download only the required waveform data
for running this example from IRIS/EarthScope.

Small adjustmentsa are made to the detect and locate scripts to reflect
the different archive structure used for the downloaded data, and the
different channel naming in the 'official' copy. The data themselves are
identical, but the channels are Z->1, N->2, E->3, and the band code is
also different. The `onset.channel_maps` are fully utilised here (along
with the changes in the previous commit) to make this example run
identically despite the unconventional channels = win.

This saves ~ 115 MB of data transfer each time the repo is cloned, for a
cost of only ~ 5 seconds to download the data on a fast internet
connection.