:atom: Use MassDownloader for Icequake_Iceland
Switch from using waveform data shipped within the repo to using the
MassDownloader to efficiently download only the required waveform data
for running this example from IRIS/EarthScope.

Small adjustments are made to the detect/locate scripts to reflect the
changes in the archive structure used for the downloaded data.

The downloaded data have some differences to the bundled copy we were
using previously. These can be summarised as:
 - downloaded has location_code === ""; some of bundled had "01"
 - downloaded has variable tr.stats.mseed.dataquality flags; bundled was
   all set to "D"
 - downloaded has correct band codes (CH[ZNE] for SKG** data, and
   DL[ZNE] for SKR** data); bundled was all HH[ZNE]
 - downloaded SKR07 data has inverted Z and E compared to bundled data;
   this was corrected in the bundled copy (see T.S. Hudson PhD thesis).
   This has no impact on onset function calculation (given the signal
   transform applied), and therefore only affects the cut waveforms.
 - SKR01 and SKG13 are very subtly different data (having been
   re-extracted and processed by TSH); reasons for the re-processing and
   explanations for the subtly different values are undocumented, but
   make no appreciable differences to any of the outputs -- only the
   cut waveforms are flagged by `test_benchmarks.py` and are therefore
   updated after careful comparison.

This saves ~ 4MB of repo space for just a few seconds of download on a
fast connection.