:atom: Compute geometric mean of Cov_Err's; output in .event & plot
The geometric mean of the three "Covariance errors" has been found to be
very useful for filtering "real" events from artefacts. This adds
pre-computation of this quantity to the `event` object, and outputs it
in the .event file and on the event summary plot.

In order to fit it neatly in the space on the summary plot the two
magnitude lines have been merged to one (which better reflects the
relative unimportance of the magnitude r2 value).