:bug: Make day-splitting behaviour consistent within Trigger
When a multi-day QM run is processed through trigger, the treatment of
the sample that falls at midnight needs to be consistent (such that the
rare case of a trigger at midnight doesn't result in duplicate
detections on consecutive days). This was not the case - now addressed
here:
 - when running trigger, if a batch endtime is at midnight, it will be
   adjusted back by one sample. The implicit correction (using a "less
   than" condition on the endtime within `_filter_events()` is then set
   back to being "less than or equal to", matching the behaviour for the
   starttime.
 - when running locate, if the specified endtime in a call to
   `read_triggered_events` is at midnight, the condition when filtering
   the resulting events DataFrame is set to "less than" rather than
   "less than or equal to" as it is for any other endtime.

The adjustment of the "batchend" time also carries through to the
trigger summary plot, which now specifies that it runs to
YYYY-MM-DDT23:59:59, rather than running to midnight on the following
day (and xlims are also correct).

Note: because of the "less than" used previously within
`_filter_events()`, it would not have previously occurred that an event
at midnight would be triggered twice. But it leads to undocumented
unintuitive behaviour with other endtimes (in my opinion). It could have
occurred that an event triggered at midnight would be located twice,
though the cost of this is just wasted compute. An alternative solution
to what's presented here would be to just use the "less than" for the
end time in both places. As long is it's clearly documented I think
either is fine.