:bug: Fix reading triggered event files in multi-day run
As described in #132, when a multi-day locate() run had a starttime
_time_ (after midnight) later than the endtime time on some following
day, the last partial day of events would be skipped. This commit fixes
this issue by bringing the day-looping in `io.read_triggered_events` in
line with how it is achieved elsewhere; by setting the `readstart` to
midnight on the start day before beginning to add increments of 24h to
loop through the relevant days.

The issues in #132 were recreated with an adaptation of the Askja_VT-DLP
example running between 2011-10-25T23:55:00 to 2011-10-26T01:00:00 ; the
bug was confirmed prior to this commit and has been confirmed fixed by
this change.

Note that in #132 it was suggested the reading of station availability
info (for plotting the trigger summary) might also be affected, but it
has been verified here that that is not the case.