:bug: Move detection threshold plotting so always shown
Building on 4a6788c, this moves all of the detection threshold plotting
section outside the if-block that checks whether any events were
triggered. Quite useful to see the relative position of the threshold
and the coalescence trace still, I think!

Also removes a now-unnecessary 'if not None'. Had a quick look to see if
anything else might be affected, but seems not. Tested with/without an
elevated threshold and all seems good.