:atom: Add option to only plot active stations in event summary
Previously all stations in the LUT were always plotted in the locate
event summary. This commit adds the option to plot only stations that
have contributed onset functions that are used for the location. In
doing so that station is omitted from both the map/cross-sections and
the waveform gather.

This can be helpful when running over a long time period with a large
and temporally variable network, where otherwise the waveform gather can
be dominated by empty rows.