:bookmark_tabs: add interface documentation
Add documentation detailing the two available interfaces to
QuakeMigrate:

- the original script-based interface;
- and the command-line interface, which provides a config-file based
  approach to running QuakeMigrate, abstracting away the explicit Python
programming required by the script-based interface.

I have tested the whole process with the new `qmctl` command,
initialising a new project (duplicating the Iceland icequake example),
and running each stage with `qmctl run`. We might expand the options to
include a means of downloading the data from a remote data management
centre, but the scripts in the examples folders should be sufficient for
now.