:bug: Fix setter for trigger threshold method
This resolves the bug introduced in 989c498, where if the threshold
method was not set at Instantiation, the following if statements related
to the threshold method would fail.

The threshold method property/setter is modified to check the chosen
method is valid and apply the correct error message if not.

In order to ensure a value is always set for the relevant parameters
(mad_window_length, etc etc) they are now all set via kwargs.get()
irrespective of the threshold method chosen. This may cause overlap if
more methods are added in future (and is a bit inelegant), but works ok
for now.

Tested with valid and invalid method strings, with and without the
relevant threshold parameters set in the run script. Also with 'dynamic'
to ensure the FutureWarning is still printed correctly.