:bookmark_tabs: refresh doc tutorials
Refresh documentation covering tutorials for each stage of QuakeMigrate,
as well as a detailed synthetic example. The Locate tutorial needs some
further work to cover the automatic picking and magnitude estimation
add-ons, but I think for now it's better to have something than nothing
on the base documentation pages.

This supercedes PRs #92 and #119.