:bookmark_tabs: add type hinting
Add type hinting throughout package and update all docstrings to reflect
these additions.

This was also a helpful exercise in examining the behaviour of many of
the functions within the package. There were some instances where by
writing out the input and return types made it quite clear that the
function could be improved, whether by returning a 'dataclass' of sorts,
to encapsulate a collection of results, or simply that the input types
were too ambiguous.