:electron: add Ethiopia swarm example
Add another example showing results for a swarm of events at Tulu Moye,
Ethiopia. This is a basic example using the same structure as a number
of others, and was originally provided by Tim Greenfield. I have updated
it here to reflect the most recent version of QuakeMigrate and the
corresponding run scripts.