:test_tube: Add `get_data` to `run_test_examples.py`
When the Icequake_Iceland example data was removed from the repo and replaced
with the get_data script (in 6694459), it was not added to the `run_test_examples.py`
script, leading to an inconvenient failure if trying to run this immediately
after a fresh clone. (The VT_Iceland get_data script was already there). This
commit resolves this.